WIP: Move `MenuPosition` and `MenuDimensions` defs from `adapter.ts` to `constants.ts`
